test: extract gate helpers and cover optional-mode fallback
Move the three temp-fixture builders out of conformance-gate.test.ts so
the file stays under the 200-line limit after the new relative-path cases.
Add an optional-mode wrong-env test that must warn and fall back (exit 0)
instead of taking the required-mode refuse-fallthrough path.
